requirements.txt
PyMySQL>=1.1.0
requests>=2.31.0
loguru>=0.7.2
PyYAML>=6.0.1
APScheduler>=3.10.4
